Steve: various 4.2.0 fixes

+++ usr/lib/libl/ltab.c
@@ -1,2 +1,2 @@
 #include "lexlib.h"
-short *yy_llist[] = {NULL};
+int *yy_llist[] = {NULL};
